[FLINK-1201] [gelly] Added PageRank as a library method and example
